requirements.txt
anyio==4.6.2
click==8.1.7
dnspython==2.7.0
email_validator >=2.2.0
fastapi==0.115.4
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
msgpack==1.1.0
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.3.6
pydantic>=2.9.2,<3.0.0
Pygments==2.18.0
pymongo==4.10.1
python-dateutil>=2.9.0
python-multipart >=0.0.17
six==1.16.0
sniffio==1.3.1
starlette>=0.41.2,<0.42.0
strawberry-graphql[fastapi]==0.247.1
tomli==2.0.2
typing_extensions>=4.12.2
uvicorn==0.32.0